Fix PIT Mutation Testing issues with Java 17
The previous commit (7978d4e) was misleading. The JVM option
`--illegal-access` is not supported for Java >= 17.

Quoting Java Documentation:
> Many tools and libraries have been updated to avoid relying on JDK
> internals and instead use standard Java APIs that were introduced
> between JDK 8 and 17. This means the --illegal-access launcher option is
> obsolete in JDK 17. Any use of this launcher option in JDK 17, whether
> with permit, warn, debug, or deny, will have no effect other than to
> issue a warning message.

So, for Java >= 17 we will need to introduce `--add-opens` options when
building with PIT Mutation Testing.

This commit will set `MAVEN_OPTS` environment variable with the
necessary `--add-opens` to start PIT Mutation Testing when building with
Java >= 17.

They are:
* `--add-opens=java.base/java.util=ALL-UNNAMED`
* `--add-opens=java.base/java.lang.reflect=ALL-UNNAMED`
* `--add-opens=java.base/java.text=ALL-UNNAMED`
* `--add-opens=java.desktop/java.awt.font=ALL-UNNAMED

The option `--illegal-access=permit` will be used only when building
with Java 16.

The JVM child process created by PIT Mutation Testing also needs these
options. So for Java 16 we will use `--illegal-access=permit` and for
Java >= 17 we will use the same `--add-opens` listed above together with
the `--add-opens=java.base/java.lang=ALL-UNNAMED`.

How we did that?

We set GitHub Actions conditional builds for Java < 16, Java == 16 and
Java >= 17 with the `MAVEN_OPTS` environment variables as presented
before and defined the Maven profiles that will configure the PIT
Mutation Testing Maven Plugin for Java 16 and Java >= 17 also as
presented before.

PIT Mutation Testing is now configured by default but is skipped. It
will only be executed  when the Maven profile `pitest` is activated.