Fix GitHub Actions: Conditional build using major Java version
On PR #20 (e7da414) we introduced a conditional build based on the major
version number of Java. But if the version is a string we can't compare
it with a number.

see: https://docs.github.com/en/actions/learn-github-actions/expressions

note: the build for Java 7.0.121 is not running.

This commit extract the major version, that is, the first number before
any dot, so we can use it to compare versions.